+++ hss_okta_flutter/lib/src/web/stubs/hss_okta_flutter_web_stub.dart
@@ -21,7 +21,7 @@ class HssOktaFlutterWeb {
     throw UnimplementedError();
   }
 
-  Future<Map> getUserInfo({
+  Future<UserClaims> getUserInfo({
     AccessToken? accessTokenObject,
     IDToken? idTokenObject,
   }) =>