+++ hss_okta_flutter/CHANGELOG.md
@@ -47,3 +47,4 @@
 
 # 1.3.1
 * removed acces to OktaAuth JS Object
+* Changed return type of getUserInfo to UserClaims